Move TestRedirectHandler to test fixtures
COAND-911
