+++ await/src/test/java/com/adyen/checkout/await/internal/ui/DefaultAwaitDelegateTest.kt
@@ -25,7 +25,7 @@ import com.adyen.checkout.core.Environment
 import com.adyen.checkout.core.exception.ComponentException
 import com.adyen.checkout.test.LoggingExtension
 import com.adyen.checkout.test.extensions.test
-import com.adyen.checkout.ui.core.internal.test.TestRedirectHandler
+import com.adyen.checkout.ui.core.internal.TestRedirectHandler
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
