+++ qr-code/src/test/java/com/adyen/checkout/qrcode/internal/ui/DefaultQRCodeDelegateTest.kt
@@ -40,8 +40,8 @@ import com.adyen.checkout.qrcode.qrCode
 import com.adyen.checkout.test.LoggingExtension
 import com.adyen.checkout.test.extensions.test
 import com.adyen.checkout.ui.core.internal.RedirectHandler
+import com.adyen.checkout.ui.core.internal.TestRedirectHandler
 import com.adyen.checkout.ui.core.internal.exception.PermissionRequestException
-import com.adyen.checkout.ui.core.internal.test.TestRedirectHandler
 import com.adyen.checkout.ui.core.internal.util.ImageSaver
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.ExperimentalCoroutinesApi
